add sql injection protection

+++ Db/Db.cs
@@ -21,7 +21,32 @@ namespace Lab7.Database
         LinkedList<string> nameTables;
         MyConsole.Log myConsole;
 
-
+        //Защита от SQL инъекций, если символ отсутствует из разрешеннных, метод вернет false
+        private bool IsLegalQuery(string strText)
+        {
+            bool islegal = false;
+            char symbol = ' ';
+            if (strText.Length > 0)
+            {
+                char[] legalchars = @"«ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyzАБВГДЕЁЖЗИЙКЛМНОПРСТУ
+    ФХЦЧШЩЭЫЪЬЮЯабвгдеёжзийклмнопрстуфхцчшщэъьыюя01234567890., „".ToCharArray();
+                islegal = true;
+                // посимвольно проверяем пришедший string
+                for (int i = 0; i < strText.Length; i++)
+                {
+                    // если символ в строке отсутсвет в массиве разрешенных, возвращаем false
+                    if (strText.LastIndexOfAny(legalchars, i, 1) < 0)
+                    {
+                        islegal = false;
+                        symbol = strText[i];
+                        break;
+                    }
+                }
+            }
+            if (!islegal)
+                myConsole.ErrorMessage($"Error \\-_-/ Использованы недопустимые символ, Строка: {strText}, Символ: {symbol}");
+            return islegal;
+        }
 
 
         //Получение названия колонок с помощью двусвязного списка
@@ -57,7 +82,7 @@ namespace Lab7.Database
                 nameTables.AddLast(row[0].ToString());
             }
 
-            DataRowCollection nameViewDataRowCollection = GetRowsInRequest($"SELECT name FROM sys.views").Rows;
+            DataRowCollection nameViewDataRowCollection = GetRowsInRequest($"SELECT name FROM sys.views;").Rows;
 
             foreach (DataRow row in nameViewDataRowCollection)
             {
@@ -82,7 +107,13 @@ namespace Lab7.Database
         //Обновление текущей переданной по названию таблицы через запрос 
         private bool UpdateCurrentTable(string nameTable)
         {
-            DataRowCollection columnsDataRowCollection = GetRowsInRequest($"SELECT COLUMN_NAME, DATA_TYPE FROM INFORMATION_SCHEMA.COLUMNS WHERE table_name = '{nameTable}'").Rows;
+            if (!IsLegalQuery(nameTable))
+            {
+                return false;
+            }
+
+            DataRowCollection columnsDataRowCollection = GetRowsInRequest($"SELECT COLUMN_NAME, DATA_TYPE FROM INFORMATION_SCHEMA.COLUMNS WHERE table_name = '{nameTable}';").Rows;
+
             if (nameColumnsCurrentTable == null)
                 nameColumnsCurrentTable = new LinkedList<string>();
             else
@@ -99,7 +130,7 @@ namespace Lab7.Database
                 typeColumnsCurrentTable.AddLast(row[1].ToString());
             }
 
-            DataRowCollection primaryKeysDataRowCollection = GetRowsInRequest($"SELECT COLUMN_NAME FROM INFORMATION_SCHEMA.KEY_COLUMN_USAGE WHERE OBJECTPROPERTY(OBJECT_ID(CONSTRAINT_SCHEMA + '.' + QUOTENAME(CONSTRAINT_NAME)), 'IsPrimaryKey') = 1 AND TABLE_NAME = '{nameTable}'").Rows;
+            DataRowCollection primaryKeysDataRowCollection = GetRowsInRequest($"SELECT COLUMN_NAME FROM INFORMATION_SCHEMA.KEY_COLUMN_USAGE WHERE OBJECTPROPERTY(OBJECT_ID(CONSTRAINT_SCHEMA + '.' + QUOTENAME(CONSTRAINT_NAME)), 'IsPrimaryKey') = 1 AND TABLE_NAME = '{nameTable}';").Rows;
             LinkedList<string> primaryColumnsNamesCollection = new LinkedList<string>();
             foreach (DataRow row in primaryKeysDataRowCollection)
             {
@@ -171,7 +202,7 @@ namespace Lab7.Database
         public LinkedList<UsersRow> GetUsersTable()
         {
             LinkedList<UsersRow> usersNames = new LinkedList<UsersRow>();
-            DataRowCollection dataRowCollection = GetRowsInRequest("SELECT name, principal_id, type, type_desc, default_schema_name, create_date, modify_date, authentication_type_desc FROM sys.database_principals WHERE (type='S' or type = 'U')").Rows;
+            DataRowCollection dataRowCollection = GetRowsInRequest("SELECT name, principal_id, type, type_desc, default_schema_name, create_date, modify_date, authentication_type_desc FROM sys.database_principals WHERE (type='S' or type = 'U');").Rows;
             foreach (DataRow row in dataRowCollection)
             {
                 string rowFifth = "";
@@ -188,6 +219,11 @@ namespace Lab7.Database
         //Создание соединения с нашей БД
         public bool MakeConnectDb(string login, string password)
         {
+            if (!IsLegalQuery(login) || !IsLegalQuery(password))
+            {
+                return false;
+            }
+
             currentNameUser = login;//DESKTOP-0U9RJHC\MSSQLSERVERNEW
             /////МЕНЯТЬ БД ДЛЯ НОВОГО СОЕДИНЕНИЯ
             nameDatabase = "DB_Maria"; 
@@ -223,7 +259,7 @@ namespace Lab7.Database
             LinkedList<List<string>> rowsTableLinkList = new LinkedList<List<string>>();
 
             LinkedList<UsersRow> usersNames = new LinkedList<UsersRow>();
-            DataRowCollection dataRowCollection = GetRowsInRequest($"SELECT * FROM {nameTable}").Rows;
+            DataRowCollection dataRowCollection = GetRowsInRequest($"SELECT * FROM {nameTable};").Rows;
             foreach (DataRow row in dataRowCollection)
             {
                 object[] objects = row.ItemArray;
@@ -274,21 +310,33 @@ namespace Lab7.Database
         //Обновляем данные в нашей таблице, совершаем запрос нашей БД
         public void UpdateDataTable(string nameTable, string columnName, string newValue, string namePrimaryId, string primaryId)
         {
+            if (!IsLegalQuery(nameTable) || !IsLegalQuery(columnName) || !IsLegalQuery(newValue) || !IsLegalQuery(namePrimaryId) || !IsLegalQuery(primaryId))
+            {
+                return;
+            }
             MakeRequest($"UPDATE [{nameTable}] SET [{columnName}] = '{newValue}' WHERE [{namePrimaryId}] = {primaryId};");
         }
 
         //Удаляем данные в нашей таблице, совершаем запрос нашей БД
         public void DeleteDataTable(string nameTable, string namePrimaryId, string primaryId)
         {
+            if (!IsLegalQuery(nameTable) || !IsLegalQuery(namePrimaryId) || !IsLegalQuery(primaryId))
+            {
+                return;
+            }
             MakeRequest($"DELETE FROM [{nameTable}] WHERE [{namePrimaryId}] = {primaryId};");
         }
 
         //Создание нового логина пользователя, совершаем запрос к нашей БД
         public bool MakeNewLoginToUser(string name, string password, string role)
         {
-            MakeRequest($"EXEC sp_addlogin '{name}', '{password}', '{nameDatabase}'");
-            MakeRequest($"EXEC sp_adduser '{name}'");
-            MakeRequest($"EXEC sp_addrolemember '{role}', '{name}'");
+            if (!IsLegalQuery(name) || !IsLegalQuery(password) || !IsLegalQuery(role))
+            {
+                return false;
+            }
+            MakeRequest($"EXEC sp_addlogin '{name}', '{password}', '{nameDatabase}';");
+            MakeRequest($"EXEC sp_adduser '{name}';");
+            MakeRequest($"EXEC sp_addrolemember '{role}', '{name}';");
 
             return true;
         }
@@ -326,6 +374,10 @@ namespace Lab7.Database
         //Добавляем данные в нашу таблицу, совершаем запрос нашей БД
         public void InsertDataTable(string nameTable)
         {
+            if (!IsLegalQuery(nameTable))
+            {
+                return;
+            }
             string request = "";
             UpdateCurrentTable(nameTable);
 
@@ -334,7 +386,7 @@ namespace Lab7.Database
             foreach (string var in nameColumnsCurrentTable)
             {
 
-                request = $"SELECT MAX([{var.ToString()}]) FROM [{nameTable}]";
+                request = $"SELECT MAX([{var.ToString()}]) FROM [{nameTable}];";
                 //request = $"SELECT MAX([{var.ToString()}]) FROM [{nameTable}]";
                 object obj = GetRowsInRequest(request).Rows[0].Field<object>(0);
                 if (obj != null)
@@ -416,7 +468,7 @@ namespace Lab7.Database
             }
             if(insertValues.Count() > 0)
                 insertValues = insertValues.Remove(insertValues.Count() - 1, 1);
-            MakeRequest($"INSERT INTO [{nameTable}] VALUES({insertValues})");
+            MakeRequest($"INSERT INTO [{nameTable}] VALUES({insertValues});");
         }
